Clean Magit
Using woman as a man reader apparently causes an ovverride qith the
functions of Magit. Calling a command resulted in searching for a
manual entry instead. Removed the setting of woman (to an apparently
free variable ) solves the problem, though now I don't know if it's
possible to display git manpages